[Feature] added authentication

+++ routes/events.py
@@ -1,26 +1,26 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 from models.events import Event, EventUpdate
 from typing import List
 from beanie import PydanticObjectId
 from models.events import Event
 from database.connection import Database
+from auth.authenticate import authenticate
 
 event_router = APIRouter(tags=["Events"])
 
 event_database = Database(Event)
 
 
-
 # endpoint to returns all events
 @event_router.get('/', response_model=List[Event])
-async def get_all_events() -> List[Event]:
+async def get_all_events(user: str = Depends(authenticate)) -> List[Event]:
     events = await event_database.get_all()
     return events
 
 
-#get event by id
+# get event by id
 @event_router.get('/{event_id}', response_model=Event)
-async def get_single_event(event_id:int) -> Event:
+async def get_single_event(event_id: PydanticObjectId, user: str = Depends(authenticate)) -> Event:
     event = await event_database.get(event_id)
 
     if not event:
@@ -34,18 +34,31 @@ async def get_single_event(event_id:int) -> Event:
 
 # endpoint to add event
 @event_router.post('/new')
-async def create_event(event_data:Event) -> dict:
-    await event_database.save(event_data)
+async def create_event(event_data: Event, user: str = Depends(authenticate)) -> dict:
+    # adding user to event as creator
+    event_data.creator = user
 
+    event_id = await event_database.save(event_data)
     return {
-        "message" : "event added successfully."
+        "message": "event added successfully.",
+        "event_id": str(event_id)
     }
 
 
 # endpoint to update event
-@event_router.put('/edit/{event_id}')
-async def update_event(event_id:PydanticObjectId, new_event_data:EventUpdate) -> Event:
-    updated_event = event_database.update(event_id, new_event_data)
+@event_router.put('/edit/{event_id}', response_model=Event)
+async def update_event(event_id: PydanticObjectId, new_event_data: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    # getting event and verifying if event creator is user in access token
+    event = await event_database.get(event_id)
+
+    if not event or event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="operation not allowed"
+        )
+
+    # updating event
+    updated_event = await event_database.update(event_id, new_event_data)
 
     if not update_event:
         raise HTTPException(
@@ -58,7 +71,15 @@ async def update_event(event_id:PydanticObjectId, new_event_data:EventUpdate) ->
 
 # endpoint to delete event by id
 @event_router.delete('/delete/{event_id}')
-async def delete_single_event(event_id : PydanticObjectId) -> dict:
+async def delete_single_event(event_id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    # verifying whether event is created by user or not
+    event = await event_database.get(event_id)
+
+    if not event or event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="operation not allowed"
+        )
 
     deleted = await event_database.delete(event_id)
 
@@ -70,5 +91,5 @@ async def delete_single_event(event_id : PydanticObjectId) -> dict:
         )
 
     return {
-        "message" : "event deleted successfully"
+        "message": "event deleted successfully"
     }
